Little optimisation

+++ source/gameshared/q_math.c
@@ -188,14 +188,14 @@ void AngleVectors( const vec3_t angles, vec3_t forward, vec3_t right, vec3_t up 
 	float sr, sp, sy, cr, cp, cy, t;
 
 	angle = DEG2RAD( angles[YAW] );
-	sy = sin( angle );
-	cy = cos( angle );
+	sy = sinf( angle );
+	cy = cosf( angle );
 	angle = DEG2RAD( angles[PITCH] );
-	sp = sin( angle );
-	cp = cos( angle );
+	sp = sinf( angle );
+	cp = cosf( angle );
 	angle = DEG2RAD( angles[ROLL] );
-	sr = sin( angle );
-	cr = cos( angle );
+	sr = sinf( angle );
+	cr = cosf( angle );
 
 	if( forward ) {
 		forward[0] = cp * cy;
